Render in Conference Wordmarks in gt Tables

Appends a column with HTML that will render in conference wordmarks and (optionally) names for table plotting.

Usage

gt_cbb_conferences(
  data,
  conf_column,
  logo_column = "conf",
  logo_height = 20,
  include_name = FALSE,
  vertical_align = "-30%"
)

Arguments

data

Pass through your plotting data

conf_column

Indicate which column contains your team variable

logo_column

Indicates the name of the HTML column

logo_height

The height of the logo or wordmark in the HTML string

include_name

Should the conference name be included in the column? Defaults to FALSE as wordmarks are distinctive.

vertical_align

The CSS styler for vertical align; defaults to -30%.

Value

Returns data with an appended HTML column.

Details

I really hate putting conference logos and names in separate columns, and this helper function will add conference wordmarks and create the HTML code to render in wordmarks and columns in the same column for tables. If you are using gt to plot, you must use fmt_markdown({logo_column}) to render in the HTML. Set logo_column to your conf variable if you wish to overwrite your conf column with the logo HTML (recommended but not default).