Add a Team Logo and Title/Subtitle to a gt Table.

This requires wrapping the function around html and passing through to tab_header in gt. See function example.

Usage

gt_cbb_logo_title(
  title,
  subtitle,
  value = NULL,
  logo_link = NULL,
  type = c("team", "conference", "headshot"),
  title_font_size = 24,
  title_font_weight = "bold",
  title_lineheight = 0.5,
  subtitle_font_size = 16,
  subtitle_font_weight = "normal",
  subtitle_lineheight = 0.5,
  logo_height = 65
)

Arguments

title

The gt table title

subtitle

The gt table subtitle

logo_link

If you want to plot a different image, pass a valid URL link to it.

title_font_size

The size of the title font (in px).

title_font_weight

The weight of the title (normal or bold).

title_lineheight

The line-height of the title.

subtitle_font_size

The size of the subtitle font (in px).

subtitle_font_weight

The weight of the subtitle (normal or bold).

subtitle_lineheight

The line-height of the subtitle.

logo_height

The height of the logo (in px).

team

A valid team name from cbbdata.